pyproject.toml
[project]
name = "genos-agentic-pipeline"
version = "0.1.0"
description = "Genos 多智能体基因组分析系统 - 基于 Genos 基础模型的端到端变异解读与临床报告生成"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = [
    "genomics",
    "genos",
    "multi-agent",
    "variant-interpretation",
    "bioinformatics"
]

dependencies = [
    # 核心依赖
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",

    # 日志与进度条
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
# 序列处理（可选，Windows 用户可能需要预编译包）
sequence = [
    "pyfaidx>=0.7.0",
    "pysam>=0.21.0",
]

# 开发工具
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
genos-pipeline = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
